requirements.txt
amqp==5.0.9
apiclient==1.0.4
appdirs==1.4.4
arrow==1.1.1
asgiref==3.4.1
astroid
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
binaryornot==0.4.4
build==0.6.0.post1
cachetools==4.2.2
celery==5.2.3
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cookiecutter==1.7.3
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
distlib==0.3.2
Django==3.2.7
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-filter==2.4.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
drf-spectacular==0.18.2
filelock==3.0.12
Flask==2.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==5.3.4
ipython==7.25.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jinja2-time==0.2.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.3.1
kombu==5.2.3
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
matplotlib==3.3.3
matplotlib-inline==0.1.2
mccabe==0.6.1
ndjson==0.3.1
nltk==3.6.7
numpy==1.19.3
oauth2client==4.1.2
oauthlib==3.1.1
packaging==20.9
pandas==1.1.4
parso==0.8.2
pep517==0.11.0
pickleshare==0.7.5
Pillow==8.0.1
pip-review==1.1.0
pipenv==2020.6.2
poyo==0.5.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.5.3
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
python-crontab==2.6.0
python-dateutil==2.8.1
python-slugify==5.0.2
pytz==2021.3
PyYAML==5.4.1
pyzmq==19.0.2
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
tzdata==2021.5
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
watchdog==2.1.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.7.0